requirements.txt
# Data collection
aiohttp>=3.9
websockets>=12.0
redis>=5.0

# Data storage & processing
pandas>=2.0
pyarrow>=14.0
numpy>=1.26

# Schema validation
pydantic>=2.0
sortedcontainers>=2.4

# Secret encryption at rest (Fernet — AES-128-CBC + HMAC-SHA256)
cryptography>=42.0

# PostgreSQL (production database)
psycopg2-binary>=2.9

# Monitoring
prometheus-client>=0.20

# Logging & config
loguru>=0.7
python-dotenv>=1.0

# API server
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9

# AI agents
anthropic>=0.25

# Billing
stripe>=7.0

# HTTP (for gap filler REST calls and website agents)
requests>=2.31
httpx>=0.27

# Testing
pytest>=7.4
pytest-asyncio>=0.23
